+++ scm/cli/main/src/api/mod.rs
@@ -6,6 +6,6 @@ mod traits;
 mod types;
 
 pub use error::CliError;
-pub use noop::{NoopCliRunner, NoopValidator};
-pub use traits::{CliRunner, Validator};
+pub use noop::{NoopCliCommand, NoopCliRunner, NoopValidator};
+pub use traits::{CliCommand, CliRunner, Validator};
 pub use types::{CliArgs, CliOutput};