+++ scm/cli/main/src/api/noop/mod.rs
@@ -1,5 +1,7 @@
+mod noop_cli_command;
 mod noop_cli_runner;
 mod noop_validator;
 
+pub use noop_cli_command::NoopCliCommand;
 pub use noop_cli_runner::NoopCliRunner;
 pub use noop_validator::NoopValidator;