+++ scm/cli/main/src/api/traits/mod.rs
@@ -1,5 +1,7 @@
+mod cli_command;
 mod cli_runner;
 mod validator;
 
+pub use cli_command::CliCommand;
 pub use cli_runner::CliRunner;
 pub use validator::Validator;