+++ scm/cli/main/src/core/noop/mod.rs
@@ -1,2 +1,3 @@
+mod noop_cli_command_impl;
 mod noop_cli_runner_impl;
 mod noop_validator_impl;