+++ scm/cli/main/src/saf/cli/mod.rs
@@ -1,0 +1,5 @@
+mod cli_command_svc;
+mod cli_runner_svc;
+
+pub use cli_command_svc::*;
+pub use cli_runner_svc::*;